requirements.txt
pip
pyyaml
pytest